feat: upgrade DatabaseHelper version to 2 and add onUpgrade schema migration

+++ app/src/main/java/com/example/trackflow/DatabaseHelper.java
@@ -7,7 +7,7 @@ import android.database.sqlite.SQLiteOpenHelper;
 public class DatabaseHelper extends SQLiteOpenHelper {
 
     private static String DATABASE_NAME = "trackflow_db";
-    private static final int DATABASE_VERSION = 1;
+    private static final int DATABASE_VERSION = 2;
 
     private static final String SQL_CREATE_TABLE_ACTIVITY = String.format(
             "CREATE TABLE %s"
@@ -15,13 +15,15 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                     + " %s TEXT NOT NULL,"
                     + " %s TEXT NOT NULL,"
                     + " %s TEXT NOT NULL,"
-                    + " %s TEXT NOT NULL)",
+                    + " %s TEXT NOT NULL,"
+                    + " %s TEXT)",
             DatabaseContract.ActivityColumns.TABLE_NAME,
             DatabaseContract.ActivityColumns._ID,
             DatabaseContract.ActivityColumns.COLUMN_TITLE,
             DatabaseContract.ActivityColumns.COLUMN_DISTANCE,
             DatabaseContract.ActivityColumns.COLUMN_DURATION,
-            DatabaseContract.ActivityColumns.COLUMN_DATE
+            DatabaseContract.ActivityColumns.COLUMN_DATE,
+            DatabaseContract.ActivityColumns.COLUMN_PATH
     );
 
     public DatabaseHelper(Context context) {
@@ -35,7 +37,12 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + DatabaseContract.ActivityColumns.TABLE_NAME);
-        onCreate(db);
+        if (oldVersion < 2) {
+            db.execSQL("ALTER TABLE " + DatabaseContract.ActivityColumns.TABLE_NAME
+                    + " ADD COLUMN " + DatabaseContract.ActivityColumns.COLUMN_PATH + " TEXT");
+        } else {
+            db.execSQL("DROP TABLE IF EXISTS " + DatabaseContract.ActivityColumns.TABLE_NAME);
+            onCreate(db);
+        }
     }
 }
